+++ packages/create-wp-project/src/migrations/index.js
@@ -55,6 +55,8 @@ import * as migration_0_4_0 from "./0.4.0.js";
 import * as migration_1_0_0 from "./1.0.0.js";
 import * as migration_2_0_0 from "./2.0.0.js";
 import * as migration_2_1_0 from "./2.1.0.js";
+import * as migration_2_2_0 from "./2.2.0.js";
+import * as migration_2_3_0 from "./2.3.0.js";
 
 /* -------------------------------------------------------------------- */
 /* Migration catalog                                                     */
@@ -73,6 +75,8 @@ const MIGRATIONS = [
   migration_1_0_0,
   migration_2_0_0,
   migration_2_1_0,
+  migration_2_2_0,
+  migration_2_3_0,
 ];
 
 /** Schema migrations run before version migrations (newest kit schema first). */